requirements.txt
decouple==0.0.7
Django==4.1.9
django-admin==2.0.2
django-cors-headers==4.3.1
django-excel-response2==3.0.5
django-extensions==3.2.3
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.0.0
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
psycopg2-binary==2.9.6
PyJWT==1.7.1
gunicorn == 21.2.0